Refactoring to blocks
